change auth-form layout
