+++ frontend/src/app/(auth)/layout.tsx
@@ -1,0 +1,7 @@
+export default function AuthLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="flex flex-col items-center justify-center bg-background">
+        {children}
+    </div>
+    );
+}
