+++ frontend/src/app/layout.tsx
@@ -1,8 +1,6 @@
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
-import { Navigation } from '@/components/layout/navigation'
-import Footer from '@/components/layout/footer'
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -19,11 +17,9 @@ export default function RootLayout({
   return (
    <html lang="en" className="dark">
       <body className="min-h-screen bg-background text-foreground">
-        <Navigation />
         <main>
           {children}
         </main>
-        <Footer />
       </body>
     </html>
   )
